Harden all endpoints

+++ server/index.js
@@ -21,29 +21,47 @@ app.get("/message", cors(corsOptions), async (req, res) => {
 });
 
 //get car by id
-app.get("/cars/:id", cors(corsOptions), async (req, res) => {
-  try {
-    const carId = req.params.id;
+app.get(
+  "/cars/:id",
+  cors(corsOptions),
+  param("id").isNumeric(),
+  async (req, res) => {
+    try {
+      //validate
+      const errors = validationResult(req);
+      if (!errors.isEmpty()) {
+        return res.status(400).json({ erros: errors.array() });
+      }
+      const carId = req.params.id;
 
-    const [car] = await promisePool.query(
-      "select * from car where car_id = ?",
-      [carId]
-    );
-    res.send(car[0]);
-  } catch (error) {
-    console.log(error);
+      const [car] = await promisePool.query(
+        "select * from car where car_id = ?",
+        [carId]
+      );
+      car[0]
+        ? res.send(car[0])
+        : res.status(404).send({ message: "Not found." });
+    } catch (error) {
+      res.status(404).send(error);
+    }
   }
-});
+);
 
 //get car by make
 app.get("/cars", cors(corsOptions), async (req, res) => {
   try {
+    //validate
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ erros: errors.array() });
+    }
+
     const make = req.query.make;
     const [car] = await promisePool.query(
       "select * from car where make = ?",
       make
     );
-    res.send(car);
+    car ? res.send(car) : res.status(404).send({ message: "Not found." });
   } catch (error) {
     res.status(404);
   }
@@ -52,13 +70,18 @@ app.get("/cars", cors(corsOptions), async (req, res) => {
 //post car
 app.post("/cars/", cors(corsOptions), async (req, res) => {
   try {
+    //validate
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ erros: errors.array() });
+    }
+
     const car = req.body;
     console.log(req.body);
     const [result] = await promisePool.execute(
       "insert into car (make, model, color, price) values (?, ?, ?, ?)",
       [car.make, car.model, car.color, car.price]
     );
-    console.log(result);
 
     const newCarId = result.insertId;
     const [newCar] = await promisePool.query(
@@ -66,7 +89,7 @@ app.post("/cars/", cors(corsOptions), async (req, res) => {
       newCarId
     );
 
-    res.send(newCar[0]);
+    car ? res.send(newCar[0]) : res.status(404).send({ message: "Not found." });
   } catch (error) {
     res.status(404);
   }
@@ -75,6 +98,12 @@ app.post("/cars/", cors(corsOptions), async (req, res) => {
 //put car
 app.put("/cars/:id", cors(corsOptions), async (req, res) => {
   try {
+    //validate
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ erros: errors.array() });
+    }
+
     const carId = req.params.id;
     const newCar = req.body;
 
@@ -82,7 +111,9 @@ app.put("/cars/:id", cors(corsOptions), async (req, res) => {
       "update car set make = ?, model = ?, color = ?, price = ? where car_id = ?",
       [newCar.make, newCar.model, newCar.color, newCar.price, carId]
     );
-    res.send({ "message ": result.info });
+    result
+      ? res.send({ "message ": result.info })
+      : res.status(404).send({ message: "Not found." });
   } catch (error) {
     res.status(404);
   }
@@ -91,12 +122,18 @@ app.put("/cars/:id", cors(corsOptions), async (req, res) => {
 //delete car
 app.delete("/cars/:id", cors(corsOptions), async (req, res) => {
   try {
+    //validate
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ erros: errors.array() });
+    }
+
     const carId = req.params.id;
     const [result] = await promisePool.execute(
       "delete from car where car_id = ?",
       [carId]
     );
-    res.send(result);
+    result ? res.send(result) : res.status(404).send({ message: "Not found." });
   } catch (error) {
     res.status(404);
   }
